+++ Scroller/useThemeScroller.js
@@ -1,5 +1,6 @@
 import {adaptEvent, forward, handle} from '@enact/core/handle';
 import {is} from '@enact/core/keymap';
+import platform from '@enact/core/platform';
 import Spotlight from '@enact/spotlight';
 import {getRect} from '@enact/spotlight/src/utils';
 import ri from '@enact/ui/resolution';
@@ -271,13 +272,14 @@ const useSpottable = (props, instances) => {
 
 		const
 			{rtl} = props,
+			negativeCoordinate = rtl && (platform.ios || platform.safari || platform.chrome >= 85) ? -1 : 1,
 			{clientWidth} = scrollContentHandle.current.scrollBounds,
 			rtlDirection = rtl ? -1 : 1,
 			{left: containerLeft} = scrollContentNode.getBoundingClientRect(),
 			scrollLastPosition = scrollPosition ? scrollPosition : scrollContentHandle.current.scrollPos.left,
-			currentScrollLeft = rtl ? (scrollContentHandle.current.scrollBounds.maxLeft - scrollLastPosition) : scrollLastPosition,
+			currentScrollLeft = rtl && negativeCoordinate === 1 ? (scrollContentHandle.current.scrollBounds.maxLeft - scrollLastPosition) : scrollLastPosition,
 			// calculation based on client position
-			newItemLeft = scrollContentNode.scrollLeft + (itemLeft - containerLeft);
+			newItemLeft = negativeCoordinate * scrollContentNode.scrollLeft + (itemLeft - containerLeft);
 		let nextScrollLeft = scrollContentHandle.current.scrollPos.left;
 
 		if (newItemLeft + itemWidth > (clientWidth + currentScrollLeft) && itemWidth < clientWidth) {